requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.1.0
chardet==3.0.4
idna==2.8
multidict==4.5.2
numpy==1.16.4
pandas==0.24.2
pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
tqdm==4.32.1
yarl==1.3.0
